requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn==1.3.2
hopsworks==4.2.*
pyarrow
python-dotenv
joblib
prefect
obspy>=1.4.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
# Optional: For model explainability (compatible with Python 3.11, numpy<2.0, scikit-learn==1.3.2)
# shap==0.43.0  # Compatible with numpy<2.0 and scikit-learn 1.3.2
# lime==0.2.0.1  # Compatible with Python 3.11 and scikit-learn 1.3.2

